[Add] ECoreNetto.Reporting and move reporting classes
