+++ ECoreNetto.Reporting/Generators/IHtmlReportGenerator.cs
@@ -1,5 +1,5 @@
 ﻿// -------------------------------------------------------------------------------------------------
-// <copyright file="IXlReportGenerator.cs" company="Starion Group S.A">
+// <copyright file="IHtmlReportGenerator.cs" company="Starion Group S.A">
 // 
 //   Copyright 2017-2024 Starion Group S.A.
 // 
@@ -18,13 +18,13 @@
 // </copyright>
 // ------------------------------------------------------------------------------------------------
 
-namespace ECoreNetto.Tools.Generators
+namespace ECoreNetto.Reporting.Generators
 {
     /// <summary>
-    /// The purpose of the <see cref="IXlReportGenerator"/> is to generate reports of an
+    /// The purpose of the <see cref="IHtmlReportGenerator"/> is to generate an HTML report of an
     /// Ecore Model
     /// </summary>
-    public interface IXlReportGenerator : IReportGenerator
+    public interface IHtmlReportGenerator : IReportGenerator
     {
     }
 }
