+++ ECoreNetto.Reporting/Generators/MarkdownReportGenerator.cs
@@ -1,5 +1,5 @@
 ﻿// -------------------------------------------------------------------------------------------------
-// <copyright file="HtmlReportGenerator.cs" company="Starion Group S.A">
+// <copyright file="MarkdownReportGenerator.cs" company="Starion Group S.A">
 // 
 //   Copyright 2017-2024 Starion Group S.A.
 // 
@@ -18,7 +18,7 @@
 // </copyright>
 // ------------------------------------------------------------------------------------------------
 
-namespace ECoreNetto.Tools.Generators
+namespace ECoreNetto.Reporting.Generators
 {
     using System;
     using System.Diagnostics;
@@ -28,15 +28,15 @@ namespace ECoreNetto.Tools.Generators
     using Microsoft.Extensions.Logging.Abstractions;
 
     /// <summary>
-    /// The purpose of the <see cref="HtmlReportGenerator"/> is to generate an HTML report of an
+    /// The purpose of the <see cref="MarkdownReportGenerator"/> is to generate a Markdown report of an
     /// Ecore Model
     /// </summary>
-    public class HtmlReportGenerator : HandleBarsReportGenerator, IHtmlReportGenerator
+    public class MarkdownReportGenerator : HandleBarsReportGenerator, IMarkdownReportGenerator
     {
         /// <summary>
         /// The <see cref="ILogger"/> used to log
         /// </summary>
-        private readonly ILogger<HtmlReportGenerator> logger;
+        private readonly ILogger<MarkdownReportGenerator> logger;
 
         /// <summary>
         /// Initializes a new instance of the <see cref="HtmlReportGenerator"/> class.
@@ -44,9 +44,9 @@ namespace ECoreNetto.Tools.Generators
         /// <param name="loggerFactory">
         /// The (injected) <see cref="ILoggerFactory"/> used to set up logging
         /// </param>
-        public HtmlReportGenerator(ILoggerFactory loggerFactory = null) : base(loggerFactory)
+        public MarkdownReportGenerator(ILoggerFactory loggerFactory = null) : base(loggerFactory)
         {
-            this.logger = loggerFactory == null ? NullLogger<HtmlReportGenerator>.Instance : loggerFactory.CreateLogger<HtmlReportGenerator>();
+            this.logger = loggerFactory == null ? NullLogger<MarkdownReportGenerator>.Instance : loggerFactory.CreateLogger<MarkdownReportGenerator>();
         }
 
         /// <summary>
@@ -57,11 +57,11 @@ namespace ECoreNetto.Tools.Generators
         /// </returns>
         public string QueryReportType()
         {
-            return "html";
+            return "markdown";
         }
 
         /// <summary>
-        /// Generates a table that contains all classes, attributes and their documentation
+        /// Generates a Markdown document with a table that contains all classes, attributes and their documentation
         /// </summary>
         /// <param name="modelPath">
         /// the path to the Ecore model of which the report is to be generated.
@@ -71,15 +71,21 @@ namespace ECoreNetto.Tools.Generators
         /// </param>
         public void GenerateReport(FileInfo modelPath, FileInfo outputPath)
         {
-            ArgumentNullException.ThrowIfNull(modelPath);
+            if (modelPath == null)
+            {
+                throw new ArgumentNullException(nameof(modelPath));
+            }
 
-            ArgumentNullException.ThrowIfNull(outputPath);
+            if (outputPath == null)
+            {
+                throw new ArgumentNullException(nameof(outputPath));
+            }
 
             var sw = Stopwatch.StartNew();
 
-            this.logger.LogInformation("Start Generating Html report tables");
+            this.logger.LogInformation("Start Generating Markdown report");
 
-            var template = this.Templates["ecore-to-html-docs"];
+            var template = this.Templates["ecore-to-markdown-docs"];
 
             var rootPackage = this.LoadRootPackage(modelPath);
 
@@ -95,9 +101,9 @@ namespace ECoreNetto.Tools.Generators
             using var writer = outputPath.CreateText();
             writer.Write(generatedHtml);
 
-            this.logger.LogInformation("Generated HTML report in {ElapsedTime} [ms]", sw.ElapsedMilliseconds);
+            this.logger.LogInformation("Generated Markdown report in {ElapsedTime} [ms]", sw.ElapsedMilliseconds);
         }
-
+        
         /// <summary>
         /// Verifies whether the extension of the <paramref name="outputPath"/> is valid or not
         /// </summary>
@@ -110,13 +116,13 @@ namespace ECoreNetto.Tools.Generators
         /// </returns>
         public override Tuple<bool, string> IsValidReportExtension(FileInfo outputPath)
         {
-            if (outputPath.Extension == ".html")
+            if (outputPath.Extension == ".md")
             {
-                return new Tuple<bool, string>(true, ".html is a supported report extension");
+                return new Tuple<bool, string>(true, ".md is a supported report extension");
             }
 
             return new Tuple<bool, string>(false,
-                $"The Extension of the output file '{outputPath.Extension}' is not supported. Supported extensions is '.html'");
+                $"The Extension of the output file '{outputPath.Extension}' is not supported. Supported extensions is '.md'");
         }
 
         /// <summary>
@@ -124,7 +130,7 @@ namespace ECoreNetto.Tools.Generators
         /// </summary>
         protected override void RegisterTemplates()
         {
-            this.RegisterEmbeddedTemplate("ecore-to-html-docs");
+            this.RegisterEmbeddedTemplate("ecore-to-markdown-docs");
         }
     }
 }
