+++ ECoreNetto.Reporting/Templates/ecore-to-html-docs.hbs
