+++ ECoreNetto.Reporting/Templates/ecore-to-markdown-docs.hbs
