+++ ECoreNetto.Tools.Tests/ECoreNetto.Tools.Tests.csproj
@@ -13,7 +13,7 @@
         <PackageReference Include="Moq" Version="4.20.72" />
         <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.11.1" />
         <PackageReference Include="NUnit" Version="4.2.2" />
-        <PackageReference Include="NUnit.Console" Version="3.18.1" />
+        <PackageReference Include="NUnit.Console" Version="3.18.2" />
         <PackageReference Include="NUnit3TestAdapter" Version="4.6.0">
             <PrivateAssets>all</PrivateAssets>
             <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
